requirements.txt
fastapi
uvicorn
tarantool
passlib
PyJWT
asynctnt
httpx
pytest
pytest-asyncio